permalink slugs based on hashids
split the page with last 100 messages with live update feed (unimplemented)
from the one that renders a "look at history based on slug" page

slugs are based on the message timestamp but converted to short strings
with the hashids library. old style date time strings can be used too,
but parsing of a date-time string depends on the local time zone too.

so hashids and a UTC timestamp are a better slug.

back/forward pagination works on the historic page