fixup timestamp possibly being `null`
there was a small chance that timestamp was null, and the function using
it was still called before `redirect_to` destroyed the current
component. typescript was legitimately complaining about that.

the fix is to throw an exception when the permalink did not parse,
so the function does not continue,
but make the function an async one, so that only the promise is rejected,
not the whole <script> block.